moved to unit test
